requirements.txt
redis
celery
grpcio
grpcio-tools